Server-sent events

+++ apps/api/src/resources/pay-periods/payPeriods.service.ts
@@ -119,7 +119,9 @@ export class PayPeriodsService extends AvailableForUserCompany {
     }
 
     async delete(ids: number[]): Promise<void> {
-        await this.repository.delete(ids);
+        if (ids.length) {
+            await this.repository.delete(ids);
+        }
     }
 
     async findCurrent(
